backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "dmitrii",email = "mathmtit@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "django-clerk (>=0.1.15,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
